Updated Weapon System
Changed weapon system:

Weaponsystem is a class attached to the player that handles collisions between players and weapons, determines which weapon the player holds, etc.

Since weapons will have alot in common yet different functionality, I felt it best to use inheritance.
Weapon is a superclass.
Shotgun and sword inherit from weapon.

Added code to NonVrCharacterController:
I have a method called LetGoOfWeapon in the weaponsystem class that drops the weapon from the players hands. I wanted to call this method when the player uses a car. I called this method inside the OnTriggerEnter method in the NonVrCharacterController class.